Rename frontend/ to nms/
