+++ cnaas/urls.py
@@ -18,6 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('cnaas/', include('cnaas.frontend.urls')),
+    path('nms/', include('cnaas.nms.urls')),
     path('nac/', include('cnaas.nac.urls')),
 ]
